restore original functionalities
