+++ packages/subgraph/src/dao/utils.ts
@@ -84,95 +84,3 @@ export function decodeWithdrawParams(data: ByteArray): WithdrawParams {
   withdrawParams.reference = referenceBytes.toString();
   return withdrawParams;
 }
-
-function createErc20VotingPackage(who: Address, daoId: string): void {
-  let packageEntity = ERC20VotingPlugin.load(who.toHexString());
-  if (!packageEntity) {
-    packageEntity = new ERC20VotingPlugin(who.toHexString());
-    let contract = ERC20VotingContract.bind(who);
-    let relativeSupportThresholdPct = contract.try_relativeSupportThresholdPct();
-    let totalSupportThresholdPct = contract.try_totalSupportThresholdPct();
-    let minDuration = contract.try_minDuration();
-    let token = contract.try_getVotingToken();
-
-    packageEntity.relativeSupportThresholdPct = relativeSupportThresholdPct.reverted
-      ? null
-      : relativeSupportThresholdPct.value;
-    packageEntity.totalSupportThresholdPct = totalSupportThresholdPct.reverted
-      ? null
-      : totalSupportThresholdPct.value;
-    packageEntity.minDuration = minDuration.reverted ? null : minDuration.value;
-
-    packageEntity.token = token.reverted ? null : handleERC20Token(token.value);
-
-    // Create template
-    let context = new DataSourceContext();
-    context.setString('daoAddress', daoId);
-    ERC20Voting.createWithContext(who, context);
-
-    packageEntity.save();
-  }
-}
-
-function createAllowlistVotingPackage(who: Address, daoId: string): void {
-  let packageEntity = AllowlistPlugin.load(who.toHexString());
-  if (!packageEntity) {
-    packageEntity = new AllowlistPlugin(who.toHexString());
-    let contract = AllowlistVotingContract.bind(who);
-    let relativeSupportThresholdPct = contract.try_relativeSupportThresholdPct();
-    let totalSupportThresholdPct = contract.try_totalSupportThresholdPct();
-    let minDuration = contract.try_minDuration();
-
-    packageEntity.relativeSupportThresholdPct = relativeSupportThresholdPct.reverted
-      ? null
-      : relativeSupportThresholdPct.value;
-    packageEntity.totalSupportThresholdPct = totalSupportThresholdPct.reverted
-      ? null
-      : totalSupportThresholdPct.value;
-    packageEntity.minDuration = minDuration.reverted ? null : minDuration.value;
-
-    // Create template
-    let context = new DataSourceContext();
-    context.setString('daoAddress', daoId);
-    AllowlistVoting.createWithContext(who, context);
-
-    packageEntity.save();
-  }
-}
-
-export function addPackage(daoId: string, who: Address): void {
-  // package
-  // TODO: rethink this once the market place is ready
-  let contract = ERC165Contract.bind(who);
-
-  let ERC20VotingInterfaceSuppoted = supportsInterface(
-    contract,
-    ERC20_VOTING_INTERFACE
-  );
-  let allowlistInterfaceSuppoted = supportsInterface(
-    contract,
-    ALLOWLIST_VOTING_INTERFACE
-  );
-
-  if (ERC20VotingInterfaceSuppoted) {
-    createErc20VotingPackage(who, daoId);
-  } else if (allowlistInterfaceSuppoted) {
-    createAllowlistVotingPackage(who, daoId);
-  }
-
-  if (ERC20VotingInterfaceSuppoted || allowlistInterfaceSuppoted) {
-    let daoPluginEntityId = daoId + '_' + who.toHexString();
-    let daoPluginEntity = new DaoPlugin(daoPluginEntityId);
-    daoPluginEntity.plugin = who.toHexString();
-    daoPluginEntity.dao = daoId;
-    daoPluginEntity.save();
-  }
-}
-
-export function removePackage(daoId: string, who: string): void {
-  let daoPluginEntityId = daoId + '_' + who;
-  let daoPluginEntity = DaoPlugin.load(daoPluginEntityId);
-  if (daoPluginEntity) {
-    store.remove('DaoPlugin', daoPluginEntityId);
-  }
-}
