+++ packages/subgraph/tests/dao/permission-manager.test.ts
@@ -1,47 +1,27 @@
 import {assert, clearStore, test} from 'matchstick-as/assembly/index';
-import {
-  Address,
-  BigInt,
-  ByteArray,
-  Bytes,
-  crypto
-} from '@graphprotocol/graph-ts';
+import {Address, ByteArray, Bytes, crypto} from '@graphprotocol/graph-ts';
 
 import {handleFrozen, handleGranted, handleRevoked} from '../../src/dao/dao';
 import {Permission, ContractPermissionId} from '../../generated/schema';
 import {
   DAO_ADDRESS,
   ADDRESS_ONE,
-  DAO_TOKEN_ADDRESS,
-  ONE_ETH,
   VOTING_ADDRESS,
   ADDRESS_TWO
 } from '../constants';
-import {createTokenCalls} from '../utils';
 import {
   createNewFrozenEvent,
   createNewGrantedEvent,
   createNewRevokedEvent,
   getEXECUTE_PERMISSION_ID,
-  getEXECUTE_PERMISSION_IDreverted,
-  getParticipationRequiredPct,
-  getSupportRequiredPct,
-  getVotingToken,
-  getVotesLength,
-  getMinDuration,
-  getSupportsInterface
+  getEXECUTE_PERMISSION_IDreverted
 } from './utils';
-
-import {
-  ERC20_VOTING_INTERFACE,
-  ALLOWLIST_VOTING_INTERFACE
-} from '../../src/utils/constants';
 
 let contractPermissionId = Bytes.fromByteArray(
   crypto.keccak256(ByteArray.fromUTF8('EXECUTE_PERMISSION'))
 );
 
-function testPackages(supportsErc20VotingInterface: boolean): void {
+test('Run dao (handleGranted) mappings with mock event', () => {
   // create event and run it's handler
   let grantedEvent = createNewGrantedEvent(
     contractPermissionId,
@@ -51,31 +31,6 @@ function testPackages(supportsErc20VotingInterface: boolean): void {
     ADDRESS_TWO,
     DAO_ADDRESS
   );
-
-  // launch calls
-  getEXECUTE_PERMISSION_ID(DAO_ADDRESS, contractPermissionId);
-  getSupportRequiredPct(VOTING_ADDRESS, BigInt.fromString(ONE_ETH));
-  getParticipationRequiredPct(VOTING_ADDRESS, BigInt.fromString(ONE_ETH));
-  getMinDuration(VOTING_ADDRESS, BigInt.fromString(ONE_ETH));
-  getVotesLength(VOTING_ADDRESS, BigInt.fromString(ONE_ETH));
-
-  if (supportsErc20VotingInterface) {
-    createTokenCalls(DAO_TOKEN_ADDRESS, 'DAO Token', 'DAOT', '6');
-    getVotingToken(VOTING_ADDRESS, DAO_TOKEN_ADDRESS);
-    getSupportsInterface(
-      VOTING_ADDRESS,
-      ERC20_VOTING_INTERFACE,
-      supportsErc20VotingInterface
-    );
-    getSupportsInterface(VOTING_ADDRESS, ALLOWLIST_VOTING_INTERFACE, false);
-  } else {
-    getSupportsInterface(
-      VOTING_ADDRESS,
-      ERC20_VOTING_INTERFACE,
-      supportsErc20VotingInterface
-    );
-    getSupportsInterface(VOTING_ADDRESS, ALLOWLIST_VOTING_INTERFACE, true);
-  }
 
   // handle event
   handleGranted(grantedEvent);
@@ -131,40 +86,7 @@ function testPackages(supportsErc20VotingInterface: boolean): void {
     permissionEntityID
   );
 
-  // governance
-  let daoPluginEntityID =
-    Address.fromString(DAO_ADDRESS).toHexString() +
-    '_' +
-    Address.fromString(VOTING_ADDRESS).toHexString();
-
-  assert.fieldEquals('DaoPlugin', daoPluginEntityID, 'id', daoPluginEntityID);
-
-  // packages
-  if (supportsErc20VotingInterface) {
-    assert.fieldEquals(
-      'ERC20VotingPlugin',
-      Address.fromString(VOTING_ADDRESS).toHexString(),
-      'id',
-      Address.fromString(VOTING_ADDRESS).toHexString()
-    );
-  } else {
-    assert.fieldEquals(
-      'AllowlistPlugin',
-      Address.fromString(VOTING_ADDRESS).toHexString(),
-      'id',
-      Address.fromString(VOTING_ADDRESS).toHexString()
-    );
-  }
-
   clearStore();
-}
-
-test('Run dao (handleGranted) mappings with mock event for ERC20 Voting', () => {
-  testPackages(true);
-});
-
-test('Run dao (handleGranted) mappings with mock event for Allowlist Voting', () => {
-  testPackages(false);
 });
 
 test('Run dao (handleGranted) mappings with reverted mocke call', () => {
